course_management_system/requirements.txt

django
djangorestframework
drf-spectacular
djangorestframework-simplejwt
python-decouple